mm, memory_hotplug: move zone/pages handling to offline stage
The current implementation accesses pages during hot-remove stage in order
to get the zone linked to this memory-range.  We use that zone for a)
check if the zone is ZONE_DEVICE and b) to shrink the zone's spanned
pages.

Accessing pages during this stage is problematic, as we might be accessing
pages that were not initialized if we did not get to online the memory
before removing it.

The only reason to check for ZONE_DEVICE in __remove_pages is to bypass
the call to release_mem_region_adjustable(), since these regions are
removed with devm_release_mem_region.

With patch#2, this is no longer a problem so we can safely call
release_mem_region_adjustable().  release_mem_region_adjustable() will
spot that the region we are trying to remove was acquired by means of
devm_request_mem_region, and will back off safely.

This allows us to remove all zone-related operations from hot-remove
stage.

Because of this, zone's spanned pages are shrinked during the offlining
stage in shrink_zone_pgdat().  It would have been great to decrease also
the spanned page for the node there, but we need them in
try_offline_node().  So we still decrease spanned pages for the node in
the hot-remove stage.

The only particularity is that now
find_smallest_section_pfn/find_biggest_section_pfn, when called from
shrink_zone_span, will now check for online sections and not valid
sections instead.  To make this work with devm/HMM code, we need to call
offline_mem_sections and online_mem_sections in that code path when we are
adding memory.

Link: http://lkml.kernel.org/r/20181127162005.15833-4-osalvador@suse.de
Signed-off-by: Oscar Salvador <osalvador@suse.de>
Cc: Dan Williams <dan.j.williams@intel.com>
Cc: David Hildenbrand <david@redhat.com>
Cc: Jerome Glisse <jglisse@redhat.com>
Cc: Jonathan Cameron <Jonathan.Cameron@huawei.com>
Cc: Michal Hocko <mhocko@suse.com>
Cc: Pavel Tatashin <pasha.tatashin@soleen.com>
Cc: "Rafael J. Wysocki" <rafael@kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>