zram: writeback throttle
On small memory systems there are lots of write IOs so if we use a flash
device as swap there would be serious flash wearout.  To overcome this
problem, system developers need to design a write limitation strategy to
guarantee flash health for the entire product life.

This patch creates a new knob "writeback_limit" on zram.  With that, if
the current writeback IO count (/sys/block/zramX/bd_stat) exceeds the
limitation, zram stops further writeback until the admin can reset the
limit.

Link: http://lkml.kernel.org/r/20181127055429.251614-8-minchan@kernel.org
Signed-off-by: Minchan Kim <minchan@kernel.org>
Reviewed-by: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>
Cc: Joey Pabalinas <joeypabalinas@gmail.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>