mm: thp: fix soft dirty for migration when split
When splitting a huge migrating PMD, we'll transfer the soft dirty bit
from the huge page to the small pages.  However we're possibly using wrong
data since when fetching the bit we're using pmd_soft_dirty() upon a
migration entry.  Fix it up.

If my understanding is correct about the problem then if without the patch
there is chance to lose some of the dirty bits in the migrating pmd pages
(on x86_64 we're fetching bit 11 which is part of swap offset instead of
bit 2) and it could potentially corrupt the memory of an userspace program
which depends on the dirty bit.

Link: http://lkml.kernel.org/r/20181206084604.17167-1-peterx@redhat.com
Signed-off-by: Peter Xu <peterx@redhat.com>
Cc: Andrea Arcangeli <aarcange@redhat.com>
Cc: "Kirill A. Shutemov" <kirill.shutemov@linux.intel.com>
Cc: Matthew Wilcox <willy@infradead.org>
Cc: Michal Hocko <mhocko@suse.com>
Cc: Dave Jiang <dave.jiang@intel.com>
Cc: "Aneesh Kumar K.V" <aneesh.kumar@linux.vnet.ibm.com>
Cc: Souptick Joarder <jrdr.linux@gmail.com>
Cc: Konstantin Khlebnikov <khlebnikov@yandex-team.ru>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>