checkstack.pl: fix for aarch64
There is actually a space after "sp," like this,

ffff2000080813c8:       a9bb7bfd        stp     x29, x30, [sp, #-80]!

Right now, checkstack.pl isn't able to print anything on aarch64, because
it won't be able to match the stating objdump line of a function due to
this missing space.  Hence, it displays every stack as zero-size.

After this patch, checkpatch.pl is able to match the start of a function's
objdump, and is then able to calculate each function's stack correctly.

Link: http://lkml.kernel.org/r/20181207195843.38528-1-cai@lca.pw
Signed-off-by: Qian Cai <cai@lca.pw>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>